Let CommonVoltageLevel extend VoltageLevel
