+++ src/main/java/edu/ie3/models/input/NodeInput.java
@@ -5,9 +5,9 @@
 */
 package edu.ie3.models.input;
 
-import edu.ie3.models.CommonVoltageLevel;
 import edu.ie3.models.OperationTime;
 import edu.ie3.models.StandardUnits;
+import edu.ie3.models.voltagelevels.CommonVoltageLevel;
 import java.util.Objects;
 import java.util.UUID;
 import javax.measure.Quantity;
